refactor: extract public methods to an interface
This sets us up to support OIDC clients in the future.
